requirements.txt
requests==2.32.3
python-dotenv==1.0.1
openai>=1.52.0
pydantic>=2.12.0